requirements.txt
opencv-python-headless==4.12.0.88
pillow==12.0.0
numpy==2.2.6
pytesseract==0.3.13